move shared stuff to package/studio-web
