+++ packages/studio-web/public/apple-touch-icon.png
