+++ packages/studio-web/public/favicon-16x16.png
