+++ packages/studio-web/public/favicon-32x32.png
