+++ packages/studio-web/public/favicon.ico
