+++ packages/studio-web/src/services/LocalStorageAppConfiguration.ts
