+++ packages/studio-web/tsconfig.json
@@ -3,6 +3,8 @@
   "include": ["**/*"],
   "compilerOptions": {
     "module": "commonjs",
+    "jsx": "react-jsx",
+    "lib": ["dom", "dom.iterable", "es2020", "webworker"],
     "noEmit": true,
     "experimentalDecorators": true
   }
