+++ web/src/entrypoint.tsx
@@ -2,6 +2,6 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
-import { main } from "./main";
+import { main } from "@foxglove/studio-web";
 
 void main();
