+++ web/webpack.config.ts
@@ -2,154 +2,13 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
-import ReactRefreshPlugin from "@pmmmwh/react-refresh-webpack-plugin";
-import SentryWebpackPlugin from "@sentry/webpack-plugin";
-import { CleanWebpackPlugin } from "clean-webpack-plugin";
-import CopyPlugin from "copy-webpack-plugin";
-import HtmlWebpackPlugin from "html-webpack-plugin";
 import path from "path";
-import { Configuration, EnvironmentPlugin, WebpackPluginInstance } from "webpack";
-import type { Configuration as WebpackDevServerConfiguration } from "webpack-dev-server";
 
-import type { WebpackArgv } from "@foxglove/studio-base/WebpackArgv";
-import { buildEnvironmentDefaults } from "@foxglove/studio-base/environment";
-import { makeConfig } from "@foxglove/studio-base/webpack";
-
-export interface WebpackConfiguration extends Configuration {
-  devServer?: WebpackDevServerConfiguration;
-}
-
-export type ConfigParams = {
-  contextPath: string;
-  outputPath: string;
-};
-
-export const devServerConfig = (params: ConfigParams): WebpackConfiguration => ({
-  // Use empty entry to avoid webpack default fallback to /src
-  entry: {},
-
-  // Output path must be specified here for HtmlWebpackPlugin within render config to work
-  output: {
-    publicPath: "",
-    path: params.outputPath,
-  },
-
-  devServer: {
-    static: {
-      directory: params.outputPath,
-    },
-    hot: true,
-    // The problem and solution are described at <https://github.com/webpack/webpack-dev-server/issues/1604>.
-    // When running in dev mode two errors are logged to the dev console:
-    //  "Invalid Host/Origin header"
-    //  "[WDS] Disconnected!"
-    // Since we are only connecting to localhost, DNS rebinding attacks are not a concern during dev
-    allowedHosts: "all",
-  },
-
-  plugins: [new CleanWebpackPlugin()],
-});
-
-export const mainConfig =
-  (params: ConfigParams) =>
-  (env: unknown, argv: WebpackArgv): Configuration => {
-    const isDev = argv.mode === "development";
-    const isServe = argv.env?.WEBPACK_SERVE ?? false;
-    const publicPath = process.env.FOXGLOVE_PUBLIC_PATH ?? "";
-
-    const allowUnusedVariables = isDev;
-
-    const plugins: WebpackPluginInstance[] = [];
-
-    if (isServe) {
-      plugins.push(new ReactRefreshPlugin());
-    }
-
-    // Source map upload if configuration permits
-    if (
-      !isDev &&
-      process.env.SENTRY_AUTH_TOKEN != undefined &&
-      process.env.SENTRY_ORG != undefined &&
-      process.env.SENTRY_PROJECT != undefined
-    ) {
-      plugins.push(
-        new SentryWebpackPlugin({
-          authToken: process.env.SENTRY_AUTH_TOKEN,
-          org: process.env.SENTRY_ORG,
-          project: process.env.SENTRY_PROJECT,
-          include: params.outputPath,
-          setCommits:
-            process.env.SENTRY_REPO && process.env.SENTRY_CURRENT_COMMIT
-              ? { repo: process.env.SENTRY_REPO, commit: process.env.SENTRY_CURRENT_COMMIT }
-              : undefined,
-        }),
-      );
-    }
-
-    const appWebpackConfig = makeConfig(env, argv, { allowUnusedVariables });
-
-    const config: Configuration = {
-      name: "main",
-
-      ...appWebpackConfig,
-
-      target: "web",
-      context: params.contextPath,
-      entry: "./entrypoint.tsx",
-      devtool: isDev ? "eval-cheap-module-source-map" : "source-map",
-
-      output: {
-        publicPath: publicPath === "" ? "auto" : publicPath,
-
-        // Output filenames should include content hashes in order to cache bust when new versions are available
-        filename: isDev ? "[name].js" : "[name].[contenthash].js",
-
-        path: params.outputPath,
-      },
-
-      plugins: [
-        ...plugins,
-        ...(appWebpackConfig.plugins ?? []),
-        new EnvironmentPlugin(buildEnvironmentDefaults(argv.env?.FOXGLOVE_BACKEND ?? argv.mode)),
-        new CopyPlugin({
-          patterns: [{ from: "../public" }],
-        }),
-        new HtmlWebpackPlugin({
-          templateContent: `
-  <!doctype html>
-  <html>
-    <head>
-      <meta charset="utf-8">
-      <meta name="apple-mobile-web-app-capable" content="yes">
-      <meta property="og:title" content="Foxglove Studio"/>
-      <meta property="og:description" content="Open source visualization and debugging tool for robotics"/>
-      <meta property="og:type" content="website"/>
-      <meta property="og:image" content="https://foxglove.dev/images/og-image.jpeg"/>
-      <meta property="og:url" content="https://studio.foxglove.dev/"/>
-      <meta name="twitter:card" content="summary_large_image"/>
-      <meta name="twitter:site" content="@foxglovedev"/>
-      <link rel="apple-touch-icon" sizes="180x180" href="${publicPath}apple-touch-icon.png" />
-      <link rel="icon" type="image/png" sizes="32x32" href="${publicPath}favicon-32x32.png" />
-      <link rel="icon" type="image/png" sizes="16x16" href="${publicPath}favicon-16x16.png" />
-      <title>Foxglove Studio</title>
-    </head>
-    <script>
-      global = globalThis;
-    </script>
-    <body>
-      <div id="root"></div>
-    </body>
-  </html>
-  `,
-        }),
-      ],
-    };
-
-    return config;
-  };
+import { ConfigParams, devServerConfig, mainConfig } from "@foxglove/studio-web/src/webpackConfigs";
 
 const params: ConfigParams = {
   outputPath: path.resolve(__dirname, ".webpack"),
   contextPath: path.resolve(__dirname, "src"),
+  entrypoint: "./entrypoint.tsx",
 };
 export default [devServerConfig(params), mainConfig(params)];
